feat: implement CSV export for reservations by event ID with role-based access control
