+++ src/main/java/de/felixhertweck/seatreservation/reservation/service/ReservationService.java
@@ -19,14 +19,9 @@
  */
 package de.felixhertweck.seatreservation.reservation.service;
 
-import java.io.BufferedWriter;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStreamWriter;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -36,9 +31,7 @@ import jakarta.transaction.Transactional;
 
 import de.felixhertweck.seatreservation.common.exception.EventNotFoundException;
 import de.felixhertweck.seatreservation.email.EmailService;
-import de.felixhertweck.seatreservation.eventManagement.dto.ReservationExportDTO;
 import de.felixhertweck.seatreservation.eventManagement.exception.ReservationNotFoundException;
-import de.felixhertweck.seatreservation.model.entity.Roles;
 import de.felixhertweck.seatreservation.model.entity.*;
 import de.felixhertweck.seatreservation.model.repository.EventRepository;
 import de.felixhertweck.seatreservation.model.repository.EventUserAllowanceRepository;
@@ -345,64 +338,5 @@ public class ReservationService {
         LOG.infof(
                 "Sent reservation update confirmation for user %s (ID: %d) and reservation %d.",
                 currentUser.getUsername(), currentUser.getId(), reservation.id);
-    }
-
-    public byte[] exportReservationsToCsv(Long eventId, User currentUser)
-            throws EventNotFoundException, SecurityException, IOException {
-        LOG.infof(
-                "Attempting to export reservations for event ID %d by user: %s (ID: %d)",
-                eventId, currentUser.getUsername(), currentUser.getId());
-
-        Event event =
-                eventRepository
-                        .findByIdOptional(eventId)
-                        .orElseThrow(
-                                () -> {
-                                    LOG.warnf(
-                                            "Event with ID %d not found for CSV export.", eventId);
-                                    return new EventNotFoundException(
-                                            "Event with id " + eventId + " not found");
-                                });
-
-        if (!event.getManager().equals(currentUser)
-                && !currentUser.getRoles().contains(Roles.ADMIN)) {
-            LOG.warnf(
-                    "User %s (ID: %d) is not authorized to export reservations for event ID %d.",
-                    currentUser.getUsername(), currentUser.getId(), eventId);
-            throw new SecurityException(
-                    "User is not authorized to export this event's reservations");
-        }
-
-        List<Reservation> reservations = reservationRepository.findByEventId(eventId);
-
-        // Sort by seat number
-        reservations.sort(Comparator.comparing(r -> r.getSeat().getSeatNumber()));
-
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        try (BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(baos))) {
-            // CSV Header
-            writer.write("ID,Seat Number,First Name,Last Name,Reservation Date\n");
-
-            long exportIdCounter = 1;
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm");
-
-            for (Reservation reservation : reservations) {
-                ReservationExportDTO dto = new ReservationExportDTO(reservation, exportIdCounter++);
-                writer.write(
-                        String.format(
-                                "%d,%s,%s,%s,%s\n",
-                                dto.getId(),
-                                dto.getSeatNumber(),
-                                dto.getFirstName(),
-                                dto.getLastName(),
-                                dto.getReservationDate().format(formatter)));
-            }
-            writer.flush();
-        }
-
-        LOG.debugf(
-                "Successfully exported %d reservations for event ID %d to CSV by user: %s (ID: %d)",
-                reservations.size(), eventId, currentUser.getUsername(), currentUser.getId());
-        return baos.toByteArray();
     }
 }
